refactor auth config and helper functions
