+++ .gitignore
@@ -4,4 +4,6 @@ dist
 
 .vscode
 
+src/config/auth.config.js
+
 debug.log
