+++ src/components/LoginLink.js
@@ -1,5 +1,8 @@
 import React from 'react';
-import { getAuthURL } from '../config/authConfig';
+import { getAuthURL } from '../helpers/authHelpers';
+import { config } from '../config/auth.config';
 
 
-export const LoginLink = ({ message, dialog }) => (<a href={getAuthURL(dialog)}>{message}</a>);
+export const LoginLink = ({ message, dialog }) => (
+  <a href={getAuthURL(config, dialog)}>{message}</a>
+);
