+++ src/config/sample-auth.config.js
@@ -1,0 +1,6 @@
+// add your own config information here and rename this file to auth.config.js
+
+export const config = {
+  clientId: 'Your Spotify Client ID',
+  redirectUri: 'The URL to redirect back to in implicit grant auth flow.',
+};
